+++ test/spec/index.spec.ts
@@ -6,7 +6,8 @@ import { getStackTrace } from '../util/get-stack-trace';
 describe('index', () => {
 	it('sets stack trace to "Just My Code"', () => {
 		setupErrorHandling({
-			justMyCode: true
+			justMyCode: true,
+			includeJsFiles: true,
 		});
 
 		const trace = getStackTrace();
@@ -14,5 +15,15 @@ describe('index', () => {
 		expect(trace).toContain('get-stack-trace');
 		expect(trace).not.toContain('node_modules');
 		expect(trace).not.toContain('internal/');
+	});
+
+	it('allows filtering of .js extension', () => {
+		setupErrorHandling({
+			justMyCode: true
+		});
+
+		const trace = getStackTrace();
+
+		expect(trace).not.toContain('.js');
 	});
 });