+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.2]
 
 ### Fixes
+- [Issue #6] Stack traces should be filtered by .ts extension
 - Typescript Template v2
 
 ## [1.0.1]